SmartPlotter simplification
