+++ shapash/plots/__init__.py
@@ -1,0 +1,3 @@
+from .plot_compacity import plot_compacity
+
+__all__ = ["plot_compacity"]
