+++ tests/unit_tests/explainer/test_smart_plotter.py
@@ -1,8 +1,7 @@
 """
 Unit test smart plotter
 """
-import copy
-import types
+
 import unittest
 from unittest.mock import patch
 
@@ -18,8 +17,10 @@ from shapash import SmartExplainer
 from shapash.backend import ShapBackend
 from shapash.explainer.multi_decorator import MultiDecorator
 from shapash.explainer.smart_state import SmartState
+from shapash.plots.plot_line_comparison import plot_line_comparison
 from shapash.style.style_utils import get_palette
 from shapash.utils.check import check_model
+from shapash.utils.sampling import subset_sampling
 
 
 class TestSmartPlotter(unittest.TestCase):
@@ -124,7 +125,7 @@ class TestSmartPlotter(unittest.TestCase):
         assert len(list(self.smart_explainer.plot._style_dict.keys())) > 0
 
     @patch("shapash.explainer.smart_explainer.SmartExplainer.filter")
-    @patch("shapash.explainer.smart_plotter.SmartPlotter._local_pred")
+    @patch("shapash.explainer.smart_explainer.SmartExplainer._local_pred")
     def test_local_plot_1(self, local_pred, filter):
         """
         Unit test Local plot 1
@@ -191,7 +192,7 @@ class TestSmartPlotter(unittest.TestCase):
 
     @patch("shapash.explainer.smart_explainer.SmartExplainer.filter")
     @patch("shapash.explainer.smart_plotter.select_lines")
-    @patch("shapash.explainer.smart_plotter.SmartPlotter._local_pred")
+    @patch("shapash.explainer.smart_explainer.SmartExplainer._local_pred")
     def test_local_plot_4(self, local_pred, select_lines, filter):
         """
         Unit test local plot 4
@@ -237,6 +238,7 @@ class TestSmartPlotter(unittest.TestCase):
         smart_explainer_mi.mask = [mask1, mask2]
         smart_explainer_mi._case = "classification"
         smart_explainer_mi._classes = [0, 1]
+        smart_explainer_mi.inv_features_dict = {}
         smart_explainer_mi.state = MultiDecorator(SmartState())
         condition = "index == 'B'"
         output = smart_explainer_mi.plot.local_plot(query=condition)
@@ -254,7 +256,7 @@ class TestSmartPlotter(unittest.TestCase):
 
     @patch("shapash.explainer.smart_explainer.SmartExplainer.filter")
     @patch("shapash.explainer.smart_plotter.select_lines")
-    @patch("shapash.explainer.smart_plotter.SmartPlotter._local_pred")
+    @patch("shapash.explainer.smart_explainer.SmartExplainer._local_pred")
     def test_local_plot_5(self, local_pred, select_lines, filter):
         """
         Unit test local plot 5
@@ -306,6 +308,7 @@ class TestSmartPlotter(unittest.TestCase):
         smart_explainer_mi.columns_dict = {i: col for i, col in enumerate(smart_explainer_mi.x_init.columns)}
         smart_explainer_mi.mask = [mask1, mask2]
         smart_explainer_mi.masked_contributions = [mask_contrib1, mask_contrib2]
+        smart_explainer_mi.inv_features_dict = {}
         smart_explainer_mi.mask_params = {
             "features_to_hide": None,
             "threshold": None,
@@ -348,7 +351,7 @@ class TestSmartPlotter(unittest.TestCase):
 
     @patch("shapash.explainer.smart_explainer.SmartExplainer.filter")
     @patch("shapash.explainer.smart_plotter.select_lines")
-    @patch("shapash.explainer.smart_plotter.SmartPlotter._local_pred")
+    @patch("shapash.explainer.smart_explainer.SmartExplainer._local_pred")
     def test_local_plot_groups_features(self, local_pred, select_lines, filter):
         """
         Unit test local plot 6 for groups of features
@@ -516,7 +519,7 @@ class TestSmartPlotter(unittest.TestCase):
 
     @patch("shapash.explainer.smart_explainer.SmartExplainer.filter")
     @patch("shapash.explainer.smart_plotter.select_lines")
-    @patch("shapash.explainer.smart_plotter.SmartPlotter._local_pred")
+    @patch("shapash.explainer.smart_explainer.SmartExplainer._local_pred")
     def test_local_plot_multi_index(self, local_pred, select_lines, filter):
         """
         Unit test local plot multi index
@@ -562,6 +565,7 @@ class TestSmartPlotter(unittest.TestCase):
         smart_explainer_mi.columns_dict = {i: col for i, col in enumerate(smart_explainer_mi.x_init.columns)}
         smart_explainer_mi.mask = mask_multi_index
         smart_explainer_mi._case = "regression"
+        smart_explainer_mi.inv_features_dict = {}
         smart_explainer_mi.state = SmartState()
         smart_explainer_mi.y_pred = None
 
@@ -765,6 +769,7 @@ class TestSmartPlotter(unittest.TestCase):
         xpl._case = "regression"
         xpl.state = SmartState()
         xpl.y_pred = pd.DataFrame([0.46989877093, 12.749302948], columns=["pred"], index=xpl.x_init.index)
+        xpl.plot._tuning_round_digit()
         output = xpl.plot.contribution_plot(col, violin_maxf=0)
         feature_values = xpl.x_init[col].sort_values()
         contributions = xpl.contributions[col].loc[feature_values.index]
@@ -796,6 +801,7 @@ class TestSmartPlotter(unittest.TestCase):
         xpl.x_init = pd.concat([xpl.x_init] * 10, ignore_index=True)
         xpl.postprocessing_modifications = False
         xpl.y_pred = pd.concat([pd.DataFrame([0.46989877093, 12.749302948])] * 10, ignore_index=True)
+        xpl.plot._tuning_round_digit()
         output = xpl.plot.contribution_plot(col)
         new_index = xpl.x_init[col].sort_values().index
         np_hv = np.array(
@@ -1551,7 +1557,7 @@ class TestSmartPlotter(unittest.TestCase):
         xpl.proba_values = pd.DataFrame(
             data=np.array([[0.4, 0.6], [0.3, 0.7]]), columns=["class_1", "class_2"], index=xpl.x_encoded.index.values
         )
-        output = xpl.plot._local_pred("person_A", label=0)
+        output = xpl._local_pred("person_A", label=0)
         assert isinstance(output, float)
 
     def test_plot_line_comparison_1(self):
@@ -1587,8 +1593,13 @@ class TestSmartPlotter(unittest.TestCase):
                 )
             )
         expected_output = go.Figure(data=fig)
-        output = xpl.plot._plot_line_comparison(
-            ["person_A", "person_B"], var_dict, contributions, predictions=predictions, dict_features=features_dict
+        output = plot_line_comparison(
+            ["person_A", "person_B"],
+            var_dict,
+            contributions,
+            style_dict=xpl.plot._style_dict,
+            predictions=predictions,
+            dict_features=features_dict,
         )
 
         for i in range(2):
@@ -1616,10 +1627,11 @@ class TestSmartPlotter(unittest.TestCase):
         )
         predictions = [data.loc[id] for id in index]
 
-        output = xpl.plot._plot_line_comparison(
+        output = plot_line_comparison(
             index,
             var_dict,
             contributions,
+            style_dict=xpl.plot._style_dict,
             subtitle=subtitle,
             predictions=predictions,
             dict_features=xpl.inv_features_dict,
@@ -2008,12 +2020,12 @@ class TestSmartPlotter(unittest.TestCase):
         """
         smart_explainer = self.smart_explainer
 
-        df = pd.DataFrame(
+        smart_explainer.x_init = pd.DataFrame(
             {"A": [8, 90, 10, 110], "B": [4.3, 7.4, 10.2, 15.7], "C": ["C8", "C8", "C9", "C9"], "D": [1, -3, -5, -10]},
             index=[8, 9, 10, 11],
         )
 
-        output = smart_explainer.plot.correlations(df, max_features=3)
+        output = smart_explainer.plot.correlations_plot(max_features=3)
 
         assert len(output.data) == 1
         assert len(output.data[0].x) == 3
@@ -2031,7 +2043,7 @@ class TestSmartPlotter(unittest.TestCase):
             index=[8, 9, 10, 11],
         )
 
-        output = smart_explainer.plot.correlations(df, max_features=3, facet_col="C")
+        output = smart_explainer.plot.correlations_plot(df, max_features=3, facet_col="C")
 
         assert len(output.data) == 2
         assert len(output.data[0].x) == 3
@@ -2322,7 +2334,7 @@ class TestSmartPlotter(unittest.TestCase):
         model = DecisionTreeClassifier().fit(X_train, y_train)
         xpl = SmartExplainer(model=model)
         xpl.compile(x=X_train, y_target=y_train)
-        list_ind, addnote = xpl.plot._subset_sampling(max_points=10)
+        list_ind, addnote = subset_sampling(df=xpl.x_init, max_points=10)
         assert len(list_ind) == 10
         assert addnote == "Length of random Subset: 10 (33%)"
 
@@ -2335,7 +2347,7 @@ class TestSmartPlotter(unittest.TestCase):
         model = DecisionTreeClassifier().fit(X_train, y_train)
         xpl = SmartExplainer(model=model)
         xpl.compile(x=X_train, y_target=y_train)
-        list_ind, addnote = xpl.plot._subset_sampling(max_points=50)
+        list_ind, addnote = subset_sampling(df=xpl.x_init, max_points=50)
         assert len(list_ind) == 30
         assert addnote is None
 
@@ -2349,7 +2361,7 @@ class TestSmartPlotter(unittest.TestCase):
         xpl = SmartExplainer(model=model)
         xpl.compile(x=X_train, y_target=y_train)
         selection = list(range(10, 20))
-        list_ind, addnote = xpl.plot._subset_sampling(selection=selection)
+        list_ind, addnote = subset_sampling(df=xpl.x_init, selection=selection)
         assert len(list_ind) == 10
         assert addnote == "Length of user-defined Subset: 10 (33%)"
         assert list_ind == selection
@@ -2364,7 +2376,7 @@ class TestSmartPlotter(unittest.TestCase):
         xpl = SmartExplainer(model=model)
         xpl.compile(x=X_train, y_target=y_train)
         selection = list(range(10, 20))
-        list_ind, addnote = xpl.plot._subset_sampling(selection=selection, max_points=50)
+        list_ind, addnote = subset_sampling(df=xpl.x_init, selection=selection, max_points=50)
         assert len(list_ind) == 10
         assert addnote == "Length of user-defined Subset: 10 (33%)"
         assert list_ind == selection
@@ -2380,4 +2392,4 @@ class TestSmartPlotter(unittest.TestCase):
         xpl.compile(x=X_train, y_target=y_train)
         selection = np.array(list(range(10, 20)))
         with self.assertRaises(ValueError):
-            list_ind, addnote = xpl.plot._subset_sampling(selection=selection, max_points=50)
+            list_ind, addnote = subset_sampling(df=xpl.x_init, selection=selection, max_points=50)
